feat(underlay-wrapper): update UnderlayWrapper
