+++ packages/underlay-wrapper/package.json
@@ -17,6 +17,7 @@
         "react": "^16.9.0 || ^17.0.1 || ^18.0.0"
     },
     "dependencies": {
+        "@alfalab/core-components-gap": "^1.1.0",
         "classnames": "^2.3.1"
     }
 }
