+++ packages/underlay-wrapper/src/Component.screenshots.test.tsx
@@ -18,7 +18,6 @@ describe(
                     knobs: {
                         children: 'Content',
                         backgroundColor: [
-                            'none',
                             'accent',
                             'info',
                             'attention-muted',
@@ -37,8 +36,6 @@ describe(
                             'specialbg-secondary-transparent',
                             'specialbg-tertiary-transparent',
                         ],
-                        width: 100,
-                        height: 50,
                     },
                     size: { width: 400, height: 120 },
                 }),
@@ -55,8 +52,6 @@ describe(
                             'specialbg-tertiary-transparent-inverted',
                         ],
                         children: 'Content',
-                        width: 100,
-                        height: 50,
                     },
                     size: { width: 400, height: 60 },
                 }),
@@ -81,7 +76,7 @@ describe(
                 createSpriteStorybookUrl({
                     componentName: 'UnderlayWrapper',
                     knobs: {
-                        borderRadius: 8,
+                        borderRadius: 'm',
                         children: 'Content',
                         borderColor: [
                             'accent',
@@ -104,8 +99,6 @@ describe(
                         ],
                         borderSize: 2,
                         backgroundColor: 'primary',
-                        width: 100,
-                        height: 50,
                     },
                     size: { width: 400, height: 120 },
                 }),
@@ -122,11 +115,9 @@ describe(
                             'specialbg-tertiary-transparent-inverted',
                         ],
                         children: 'Content',
-                        width: 100,
-                        height: 50,
                         backgroundColor: 'none',
                         borderSize: 2,
-                        borderRadius: 8,
+                        borderRadius: 'm',
                     },
                     size: { width: 400, height: 60 },
                 }),
@@ -174,8 +165,6 @@ describe(
                             'shadow-l-hard-up',
                             'shadow-xl-hard-up',
                         ],
-                        width: 100,
-                        height: 50,
                     },
                     size: { width: 400, height: 120 },
                 }),
