--- packages/underlay-wrapper/src/__image_snapshots__/underlay-background-color-props-sprite-inverted-snap.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e30a42545f3ee5cba9dbc5550bd7acd29aca4a7d80129eca912def79a3fbaec7
-size 16346
