--- packages/underlay-wrapper/src/__image_snapshots__/underlay-background-color-props-sprite-snap.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:617c1589d05b5c1b6f8c15e33b3e2eb2352c41170707cb6bca39cf9b99af04e4
-size 75463
