--- packages/underlay-wrapper/src/__image_snapshots__/underlay-border-props-sprite-inverted-snap.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a769263c2977d79c18d50e624fae28393d5668bb006e5dbdbb7c386a63e5148d
-size 25684
