--- packages/underlay-wrapper/src/__image_snapshots__/underlay-border-props-sprite-snap.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2ed36c0f9990f2271a39c294681606aafc214be770f8530e476dc32c92ea6bb3
-size 100020
