--- packages/underlay-wrapper/src/__image_snapshots__/underlay-shadow-props-sprite-snap.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ad7c353ebfbd7da664a888f03e9e93eba44f207135188bbdd7f67cbeaa463773
-size 146615
