+++ packages/underlay-wrapper/src/__image_snapshots__/underlay-wrapper-background-color-props-sprite-inverted-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c8b7361f078bc83921b9884f07145aa932d880b84855c705a7f4d1994fc9352b
-size 601
+oid sha256:9aa6bdce1d16565ab1afbeb6f93bd89544f892d45e08b1206de4fb0c2c341f05
+size 12908
