+++ packages/underlay-wrapper/src/__image_snapshots__/underlay-wrapper-background-color-props-sprite-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c8b7361f078bc83921b9884f07145aa932d880b84855c705a7f4d1994fc9352b
-size 601
+oid sha256:48fb3126c64acd00db9798e0835b34c27b31f25611c471ea40c2b666a0eb832e
+size 55200
