+++ packages/underlay-wrapper/src/__image_snapshots__/underlay-wrapper-border-props-sprite-inverted-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c8b7361f078bc83921b9884f07145aa932d880b84855c705a7f4d1994fc9352b
-size 601
+oid sha256:379ee313267bed8734ed0df40908bfb8bc785fe34fa77736b7ded5df3dd60566
+size 20803
