+++ packages/underlay-wrapper/src/__image_snapshots__/underlay-wrapper-border-props-sprite-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c8b7361f078bc83921b9884f07145aa932d880b84855c705a7f4d1994fc9352b
-size 601
+oid sha256:fa406560312fda429f1348ee446935b618329fc8b8b42dbc5d349d9938d7c7c4
+size 82032
