+++ packages/underlay-wrapper/src/__image_snapshots__/underlay-wrapper-shadow-props-sprite-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c8b7361f078bc83921b9884f07145aa932d880b84855c705a7f4d1994fc9352b
-size 601
+oid sha256:26e693bc4b4737541eda9451e00540f058026bab2dae1484686604e88fba0c9b
+size 110580
