--- packages/underlay-wrapper/src/border.module.css
@@ -1,115 +1,0 @@
-@import '../../themes/src/default.css';
-
-.accent {
-    border-color: var(--color-light-border-accent);
-}
-
-.key {
-    border-color: var(--color-light-border-key);
-}
-
-.key-inverted {
-    border-color: var(--color-light-border-key-inverted);
-}
-
-.link {
-    border-color: var(--color-light-border-link);
-}
-
-.primary {
-    border-color: var(--color-light-border-primary);
-}
-
-.primary-inverted {
-    border-color: var(--color-light-border-primary-inverted);
-}
-
-.secondary {
-    border-color: var(--color-light-border-secondary);
-}
-
-.secondary-inverted {
-    border-color: var(--color-light-border-secondary-inverted);
-}
-
-.tertiary {
-    border-color: var(--color-light-border-tertiary);
-}
-
-.tertiary-inverted {
-    border-color: var(--color-light-border-tertiary-inverted);
-}
-
-.underline {
-    border-color: var(--color-light-border-underline);
-}
-
-.underline-inverted {
-    border-color: var(--color-light-border-underline-inverted);
-}
-
-.graphic-attention {
-    border-color: var(--color-light-graphic-attention);
-}
-
-.graphic-link {
-    border-color: var(--color-light-graphic-link);
-}
-
-.graphic-negative {
-    border-color: var(--color-light-graphic-negative);
-}
-
-.graphic-positive {
-    border-color: var(--color-light-graphic-positive);
-}
-
-.specialbg-secondary-transparent {
-    border-color: var(--color-light-specialbg-secondary-transparent);
-}
-
-.specialbg-secondary-transparent-inverted {
-    border-color: var(--color-light-specialbg-secondary-transparent-inverted);
-}
-
-.specialbg-tertiary-transparent {
-    border-color: var(--color-light-specialbg-tertiary-transparent);
-}
-
-.specialbg-tertiary-transparent-inverted {
-    border-color: var(--color-light-specialbg-tertiary-transparent-inverted);
-}
-
-.border-width-1 {
-    border-width: 1px;
-    border-style: solid;
-}
-
-.border-width-2 {
-    border-width: 2px;
-    border-style: solid;
-}
-.border-width-4 {
-    border-width: 4px;
-    border-style: solid;
-}
-
-.border-width-none {
-    border-width: none;
-}
-
-.border-radius-0 {
-    border-radius: 0;
-}
-
-.border-radius-8 {
-    border-radius: var(--border-radius-m);
-}
-
-.border-radius-16 {
-    border-radius: var(--border-radius-xl);
-}
-
-.border-radius-20 {
-    border-radius: var(--border-radius-xxl);
-}
