+++ packages/underlay-wrapper/src/docs/description.mdx
@@ -1,10 +1,6 @@
-## Размеры и отступы
-
-Компоненту можно задать любые размеры явно, либо подстроить их под вложенный контент или внешний контейнер.
-
 ## Примеры
 
-Настраиваются: тип фона, тень, радиус скругления, величина отступов, цвет и толщина обводки. 
+Настраиваются: цвет фона, тень, радиус скругления, величина отступов, цвет и толщина обводки. 
 При необходимости, можно сделать компонент кликабельным.
 
 ```jsx live
@@ -22,34 +18,48 @@ render(() => {
       alert('Клик!')
     } 
 
+    const css = `
+        div[data-test-id='test1'] {
+            cursor: pointer;
+            width: 320px;
+            height: 200px;
+        }
+        
+        div[data-test-id='test2'] {
+            background-image: url(./images/underlayWrapper.png);
+            background-repeat: no-repeat;            
+        }
+    `;
+
     return (
     <Container>
         <UnderlayWrapper 
             onClick={handleClick}
             borderSize={1}
-            borderRadius={16}
+            borderRadius='xl'
             padding={{
-                top: 16,
-                left: 16,
-                bottom: 16,
-                right: 16
+                top: 'm',
+                left:'m',
+                bottom: 'm',
+                right: 'm'
             }}
-            width={320}
-            height={200}
             backgroundColor='positive-muted'
             borderColor='graphic-positive'
+            dataTestId='test1'
         >
-                <div style={styleContent}/>
+            <div style={styleContent} />
         </UnderlayWrapper>
+        <style>{css}</style>
         <Gap size='l'/>
         <UnderlayWrapper 
             padding={{
-                top: 16,
-                left: 16,
-                bottom: 16,
-                right: 16
+                top: 'm',
+                left:'m',
+                bottom: 'm',
+                right: 'm'
             }}
-            width='max-content'>
+            dataTestId='test2'
+            >
                 <div style={styleContent}/>
         </UnderlayWrapper>
     </Container>
